Extract Filter parser from API
